addressed comments * move cose_sign to a test module * some comments for byte blobs * added issues to some TODO comments
