+++ src/cose/mod.rs
@@ -6,4 +6,5 @@ pub mod test_setup;
 mod test_nss_verify;
 mod test_nss_sign;
 mod test_decoder;
+mod cose_sign;
 mod test_cose;
